requirements.txt
python-dotenv
langchain == 0.0.352
pandas
pypdf
psycopg2-binary
pkgconfig
mysqlclient
pinecone-client
tiktoken
streamlit
huggingface_hub
transformers
jinja2